+++ src/main/java/com/bieganski/ox/BoardListener.java
@@ -3,5 +3,6 @@ package com.bieganski.ox;
 import java.util.TreeSet;
 
 public interface BoardListener {
+    //TODO remove size
     void onBoardUpdate(TreeSet<Field> fieldsWithValue, Field addedField, int size);
 }
